Break down tests into multiple files
